+++ app/src/main/kotlin/at/bitfire/davdroid/syncadapter/BaseSyncWorker.kt
@@ -9,7 +9,6 @@ import android.content.SyncResult
 import android.net.ConnectivityManager
 import android.net.wifi.WifiManager
 import android.provider.CalendarContract
-import android.provider.ContactsContract
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 import androidx.core.content.getSystemService
@@ -229,8 +228,6 @@ abstract class BaseSyncWorker(
                 AddressBookSyncer(applicationContext)
             CalendarContract.AUTHORITY ->
                 CalendarSyncer(applicationContext)
-            ContactsContract.AUTHORITY ->
-                ContactSyncer(applicationContext)
             TaskProvider.ProviderName.JtxBoard.authority ->
                 JtxSyncer(applicationContext)
             TaskProvider.ProviderName.OpenTasks.authority,
